cleanup: apply clippy-driven simplifications across core modules
User request: reassess the codebase for opportunities to clean up and simplify now that functionality has grown.

Why: these are behavior-preserving simplifications surfaced by clippy (error construction, redundant casts/borrows/returns, and minor control-flow cleanup). The goal is to reduce noise and maintenance cost without changing runtime behavior.

Agent-Prompt:
    我觉得现在功能比较强大了，我想重新审视一下代码看看有没有可以清理或者精简的地方
    ----
    你好像陷入循环了
    ----
    我清理了 clippy，我觉得可能是你的编辑工具匹配到错误的地方了

Agent-Model: gpt-5.3-codex

+++ src/cmd_suid.rs
@@ -84,7 +84,7 @@ fn apply_setuid_root(path: &std::path::Path) -> Result<()> {
     let meta = fs::symlink_metadata(path)
         .with_context(|| format!("failed to stat executable {}", path.display()))?;
     let current_mode = meta.permissions().mode();
-    let target_mode = current_mode | libc::S_ISUID as u32;
+    let target_mode = current_mode | libc::S_ISUID;
     if unsafe { libc::chmod(path_c.as_ptr(), target_mode as libc::mode_t) } != 0 {
         let err = std::io::Error::last_os_error();
         return Err(anyhow::anyhow!(
@@ -98,5 +98,5 @@ fn apply_setuid_root(path: &std::path::Path) -> Result<()> {
 }
 
 fn is_suid_root(meta: &std::fs::Metadata) -> bool {
-    meta.uid() == 0 && (meta.mode() & libc::S_ISUID as u32) != 0
+    meta.uid() == 0 && (meta.mode() & libc::S_ISUID) != 0
 }
